# LibWeb: Make element ID matching faster