--- a/Userland/Libraries/LibWeb/DOM/AccessibilityTreeNode.cpp
+++ b/Userland/Libraries/LibWeb/DOM/AccessibilityTreeNode.cpp
@@ -41,7 +41,7 @@ void AccessibilityTreeNode::serialize_tree_as_json(JsonObjectSerializer<StringBu
             MUST(object.add("name"sv, name));
             auto description = MUST(element->accessible_description(document));
             MUST(object.add("description"sv, description));
-            MUST(object.add("id"sv, element->id()));
+            MUST(object.add("id"sv, element->unique_id()));
 
             if (has_role)
                 MUST(object.add("role"sv, ARIA::role_name(*role)));