--- a/Userland/Libraries/LibWeb/SVG/SVGSymbolElement.cpp
+++ b/Userland/Libraries/LibWeb/SVG/SVGSymbolElement.cpp
@@ -41,6 +41,7 @@ void SVGSymbolElement::apply_presentational_hints(CSS::StyleProperties& style) c
 
 void SVGSymbolElement::attribute_changed(FlyString const& name, Optional<DeprecatedString> const& value)
 {
+    Base::attribute_changed(name, value);
     if (name.equals_ignoring_ascii_case(SVG::AttributeNames::viewBox))
         m_view_box = try_parse_view_box(value.value_or(""));
 }