Noticed element ID attribute lookups in a profile, so I made a stress test to isolate some of the issues and burn them down.

The main solution to all the problems is basically to cache the `id` attribute value as an `Optional<FlyString>` on `Element`. This makes checking an element's ID O(1) instead of O(n), which is very useful in a number of places.

The stress test creates a 3000-deep DOM tree and runs two selectors against a bunch of the elements.

```
BEFORE:
Single descendant: 404ms
Double descendant: 744ms

AFTER:
Single descendant: 96ms
Double descendant: 112ms
```

So basically a 4x speed-up the single-descendant selector, and 6.5x on the double-descendant one.

Here's the stress test:
```html
<script>
    document.addEventListener("DOMContentLoaded", function() {

        let insertionPoint = document.body;

        const limit = 3000;
        for (let i = 0; i < limit; ++i) {
            const e = document.createElement("div");
            e.setAttribute("id", "div" + i);
            insertionPoint.appendChild(e);
            insertionPoint = e;
        }

        {
            const before = performance.now();
            for (let i = 0; i < limit; ++i) {
                const e = document.querySelector("body #div" + i);
            }
            const elapsed = performance.now() - before;
            console.log("Single descendant: " + elapsed);
        }

        {
            const before = performance.now();
            for (let i = 0; i < limit; ++i) {
                const e = document.querySelector("body #div500 #div" + i);
            }
            const elapsed = performance.now() - before;
            console.log("Double descendant: " + elapsed);
        }

        document.body.firstChild.remove();
    });
</script>
```